fix: (platform) fix Object Status examples for stackblitz
