+++ apps/docs/src/app/platform/component-docs/platform-object-status/platform-object-status-docs.component.html
@@ -1,4 +1,4 @@
-<fd-docs-section-title id="'default'" componentName="'objectStatus'">
+<fd-docs-section-title id="default" componentName="object-status">
     Default Object Status
 </fd-docs-section-title>
 <component-example>
@@ -6,7 +6,7 @@
 </component-example>
 <code-example [exampleFiles]="platformDefaultObjectStatusHtmlType"></code-example>
 
-<fd-docs-section-title id="'Text'" componentName="'objectStatus'">
+<fd-docs-section-title id="Text" componentName="object-status">
     Object Status Text Only
 </fd-docs-section-title>
 <component-example>
@@ -16,7 +16,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title id="'iconandText'" componentName="'objectStatus'">
+<fd-docs-section-title id="iconandText" componentName="object-status">
     Object Status with Text and Icon
 </fd-docs-section-title>
 <description></description>
@@ -27,7 +27,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title id="'genericText'" componentName="'objectStatus'">
+<fd-docs-section-title id="genericText" componentName="object-status">
     Object Status with Generic Indication Colors
 </fd-docs-section-title>
 <description></description>
@@ -38,18 +38,18 @@
 
 <separator></separator>
 
-<fd-docs-section-title id="'clickable'" componentName="'objectStatus'">
+<fd-docs-section-title id="clickable" componentName="object-status">
     Clickable Object Status
 </fd-docs-section-title>
 <description>For clickable Object Status set the <code>[clickable]</code> property to <code>true</code>.</description>
 <component-example>
-    <fdp-object-status-clickable-and-icon-example></fdp-object-status-clickable-and-icon-example>
+    <fdp-platform-object-status-clickable-and-icon-example></fdp-platform-object-status-clickable-and-icon-example>
 </component-example>
 <code-example [exampleFiles]="platformObjectStatusclickableAndIconExample"></code-example>
 
 <separator></separator>
 
-<fd-docs-section-title id="'inverted'" componentName="'objectStatus'">
+<fd-docs-section-title id="inverted" componentName="object-status">
     Inverted Object Status
 </fd-docs-section-title>
 <description
@@ -63,7 +63,7 @@
 
 <separator></separator>
 
-<fd-docs-section-title id="'invertedGeneric'" componentName="'objectStatus'">
+<fd-docs-section-title id="invertedGeneric" componentName="object-status">
     Inverted Object Status with Generic Indication Colors
 </fd-docs-section-title>
 <description></description>
@@ -74,12 +74,12 @@
 
 <separator></separator>
 
-<fd-docs-section-title id="'objectStatusLarge'" componentName="'objectStatus'">
+<fd-docs-section-title id="objectStatusLarge" componentName="object-status">
     Object Status Large Design
 </fd-docs-section-title>
 <description>For Object Status in large design use <code>[large]="true"</code>.</description>
 <component-example>
-    <fdp-object-status-object-status-large-example></fdp-object-status-object-status-large-example>
+    <fdp-platform-object-status-large-example></fdp-platform-object-status-large-example>
 </component-example>
 <code-example [exampleFiles]="platformObjectStatusLargeExample"></code-example>
