+++ apps/docs/src/app/platform/component-docs/platform-object-status/platform-object-status-docs.component.ts
@@ -1,6 +1,7 @@
 import { Component, ViewEncapsulation, ChangeDetectionStrategy } from '@angular/core';
 import { ExampleFile } from '../../../documentation/core-helpers/code-example/example-file';
 
+import * as PlatformObjectStatusDefaultExampleScss from '!raw-loader!./platform-object-status-example/platform-object-status-example.component.scss';
 import * as PlatformObjectStatusDefaultExample from '!raw-loader!./platform-object-status-example/platform-object-status-example.component.html';
 import * as PlatformObjectStatusTextExample from '!raw-loader!./platform-object-status-example/platform-object-status-text-example.component.html';
 import * as PlatformObjectStatusGenericTextExample from '!raw-loader!./platform-object-status-example/platform-object-status-generic-text-example.component.html';
@@ -9,6 +10,8 @@ import * as PlatformObjectStatusClickableAndIConExample from '!raw-loader!./plat
 import * as PlatformObjectStatusInvertedTextExample from '!raw-loader!./platform-object-status-example/platform-object-status-inverted-example.component.html';
 import * as PlatformObjectStatusInvertedGenericExample from '!raw-loader!./platform-object-status-example/platform-object-status-inverted-generic-text-example.component.html';
 import * as PlatformObjectStatusLargeExample from '!raw-loader!./platform-object-status-example/platform-object-status-large-example.component.html';
+import * as PlatformObjectStatusClickableAndIconExampleTs from '!raw-loader!./platform-object-status-example/platform-object-status-clickable-and-icon-example.component';
+import * as PlatformObjectStatusLargeExampleTs from '!raw-loader!./platform-object-status-example/platform-object-status-large-example.component';
 
 @Component({
     selector: 'fdp-platform-object-status-docs',
@@ -21,7 +24,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusDefaultExample,
-            fileName: 'platform-object-status-example'
+            fileName: 'platform-object-status-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -29,7 +33,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusTextExample,
-            fileName: 'platform-object-status-text-example'
+            fileName: 'platform-object-status-text-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -37,7 +42,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusGenericTextExample,
-            fileName: 'platform-object-status-generic-text-example'
+            fileName: 'platform-object-status-generic-text-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -45,7 +51,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusTextIconExample,
-            fileName: 'platform-object-status-icon-text-example'
+            fileName: 'platform-object-status-icon-text-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -53,7 +60,14 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusClickableAndIConExample,
-            fileName: 'platform-object-status-clickable-and-icon-example'
+            fileName: 'platform-object-status-clickable-and-icon-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
+        },
+        {
+            language: 'typescript',
+            code: PlatformObjectStatusClickableAndIconExampleTs,
+            fileName: 'platform-object-status-clickable-and-icon-example',
+            component: 'PlatformObjectStatusClickableAndIconExampleComponent'
         }
     ];
 
@@ -61,7 +75,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusInvertedTextExample,
-            fileName: 'platform-object-status-inverted-example'
+            fileName: 'platform-object-status-inverted-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -69,7 +84,8 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusInvertedGenericExample,
-            fileName: 'platform-object-status-inverted-generic-text-example'
+            fileName: 'platform-object-status-inverted-generic-text-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
         }
     ];
 
@@ -77,7 +93,14 @@ export class PlatformObjectStatusDocsComponent {
         {
             language: 'html',
             code: PlatformObjectStatusLargeExample,
-            fileName: 'platform-object-status-large-example'
+            fileName: 'platform-object-status-large-example',
+            scssFileCode: PlatformObjectStatusDefaultExampleScss
+        },
+        {
+            language: 'typescript',
+            code: PlatformObjectStatusLargeExampleTs,
+            fileName: 'platform-object-status-large-example',
+            component: 'PlatformObjectStatusLargeExampleComponent'
         }
     ];
 }
