+++ apps/docs/src/app/platform/component-docs/platform-object-status/platform-object-status-example/platform-object-status-large-example.component.scss
@@ -1,0 +1,1 @@
+@import './platform-object-status-example.component';
