+++ apps/docs/src/app/platform/component-docs/platform-object-status/platform-object-status-example/platform-object-status-large-example.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fdp-platform-object-status-large-example',
+    templateUrl: './platform-object-status-large-example.component.html',
+    styleUrls: ['./platform-object-status-large-example.component.scss']
+})
+export class PlatformObjectStatusLargeExampleComponent {
+    showAlert(): void {
+        alert('clicked alert large');
+    }
+}
